Refactor, add hot testing tests
